Rephrasing some items and updating links
There were some items that my customers and I struggled to understand, and I think if they get rephrased they are more easily understood. Also, I suggest some changes in the "More Info" links, as some of them were not directing to docs related to the recommendation in question. 
This was done in collaboration with @erjosito