Update lz_checklist.en.json
Suggesting the addition of a recommendation to consider Azure AD App Proxy as a potential alternative to AVD, if users only need access to internal apps. Could be a relevant conversation if the customer is thinking of going with an AVD deployment as part of their migration. 

We initially thought this could be added to the AVD checklist, but later realized it may not a be a good fit (if we're doing an AVD review, the customer has most likely already made their decision).